requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.9
pandas>=2.1.1
python-dotenv>=1.0.0
requests>=2.31.0
huggingface-hub>=0.19.3
openai>=1.0.0
openpyxl>=3.1.2
